// packages/Common/Types/Monitor/DnsMonitor/DnsMonitorResponse.ts
import ProbeAttempt from "../../Probe/ProbeAttempt";
import DnsRecordType from "./DnsRecordType";

export interface DnsRecordResponse {
  type: DnsRecordType;
  value: string;
  ttl?: number | undefined;
}

export default interface DnsMonitorResponse {
  isOnline: boolean;
  responseTimeInMs: number;
  failureCause: string;
  records: Array<DnsRecordResponse>;
  isDnssecValid?: boolean | undefined;
  isTimeout?: boolean | undefined;
  probeAttempts?: Array<ProbeAttempt> | undefined;
  totalAttempts?: number | undefined;
}